Blod/src/requirements.txt
annotated-types==0.7.0
anyio==4.8.0
certifi==2025.1.31
charset-normalizer==3.4.1
distro==1.9.0
filetype==1.2.0
h11==0.16.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
jiter==0.8.2
joblib==1.4.2
numpy==2.2.4
openai==1.61.1
pandas==2.2.3
pydantic==2.10.6
pydantic_core==2.27.2
pyparsing==3.2.3
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2025.2
rdflib==7.1.4
requests==2.32.4
scikit-learn==1.6.1
scipy==1.15.2
six==1.17.0
sniffio==1.3.1
SPARQLWrapper==2.0.0
threadpoolctl==3.6.0
tqdm==4.67.1
typing_extensions==4.12.2
tzdata==2025.2
urllib3==2.5.0
